+++ mlrun/api/api/api.py
@@ -69,8 +69,5 @@ api_router.include_router(
 api_router.include_router(
     secrets.router, tags=["secrets"], dependencies=[Depends(deps.AuthVerifier)]
 )
-# Since grafana-proxy imitates the path schema that was set for model-endpoints, ordering between the following two
-# routers is crucial for functionality of the grafana-proxy router.
-# ** Make sure the inclusion of the grafana_proxy.router happens before the inclusion of the model_endpoints.router. **
 api_router.include_router(grafana_proxy.router, tags=["grafana", "model_endpoints"])
 api_router.include_router(model_endpoints.router, tags=["model_endpoints"])
